Fix: Add a guard clause at the beginning of ThreeLCacheCache::evict_with_candidate to check if old_pos == -1, preventing out-of-bounds access and segmentation fault when evict_predobj returns an invalid index. Enhancement: Change the type of _size from int32_t to uint64_t to support larger objects and cache sizes, preventing std::bad_array_new_length.

+++ libCacheSim/cache/eviction/3LCache/ThreeLCache.hpp
@@ -66,7 +66,7 @@ struct MetaExtra {
 class Meta {
  public:
   uint64_t _key;
-  int32_t _size;
+  uint64_t _size;
   uint64_t _past_timestamp;
   uint16_t _freq;
   MetaExtra *_extra = nullptr;
